+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - cli: Show all todos in `todo list` command
+- cli: Update json output format for event and todo commands
 
 ### Deprecated